revert changes to database notification
