--- wire-ios-data-model/Source/Notifications/SelfPostingNotification.swift
@@ -1,36 +1,0 @@
-//
-// Wire
-// Copyright (C) 2026 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-// MARK: - Self Posting Notification
-
-public protocol SelfPostingNotification {
-    static var notificationName: Notification.Name { get }
-}
-
-public extension SelfPostingNotification {
-    static var userInfoKey: String { notificationName.rawValue }
-
-    func post(in context: NotificationContext, object: AnyObject? = nil) {
-        NotificationInContext(
-            name: type(of: self).notificationName,
-            context: context,
-            object: object,
-            userInfo: [type(of: self).userInfoKey: self]
-        ).post()
-    }
-}
